requirements.txt
certifi==2021.10.8
chardet==4.0.0
idna==2.10
urllib3==1.26.19
websockets
google-generativeai==0.3.0
asyncio
fastapi
uvicorn
google-genai==0.3.0 